requirements.txt
Flask == 2.3.2
python-dotenv == 1.0.0
psycopg2-binary == 2.9.6
Flask-SQLAlchemy == 3.0.3
Flask-Migrate == 4.0.4
flask-marshmallow == 0.15.0 
Flask-JWT-Extended == 4.5.2
cloudinary == 1.33.0
itsdangerous == 2.1.2 
Flask-Mail == 0.9.1 
Flask-Cors == 3.0.10
gunicorn == 21.2.0
flask-swagger-ui == 4.11.1 
